добавлены имена игроков рядом с xp
